tags: always pass correct buffer length
In notmuch editor we was passing sizeof(buf) with buf a char* with an
unknown allocated size.

In imap, we have assumed/hardcoded LONGSTRING. That's true, but we
should not do that.

edit_msg_tags() now takes the buflen as parameter so the root caller
can pass the size of the buffer.